LayoutTests/platform/ios-sim-deprecated/fast/forms/datetime/datetime-setrangetext-expected.txt

Test setRangeText() method is not available in datetime inputs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Running tests on input with attributes: {"type":"datetime"}

element.value = '0123456789XYZ'
FAIL element.setRangeText('ABC', 0, 0) should throw an exception. Was undefined.
PASS element.value is "0123456789XYZ"
PASS successfullyParsed is true

TEST COMPLETE
